# PERF: introduce SSE2 implementations of allnan/anynan, ~30% speedup