SSE2 is pretty much universal, so start explicitly invoking it as compilers don't seem to want to use it for `allnan`/`anynan`. 

Speedup is sufficient to bring us to parity with numpy in the slow case, which means we are now strictly better than numpy here. Still need to implement for `axis != None`, but looking very promising:
```
asv compare HEAD^ HEAD -s --sort ratio --only-changed
       before           after         ratio
     [61eea5ab]       [ff4ee4b4]
     <ppc64le>                  
-        262±10μs          224±9μs     0.86  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'F', None, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-        778±40μs          604±9μs     0.78  reduce.TimeAnyNan2D.time_anynan('float64', (1000, 1000), 'F', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        790±30μs         605±90μs     0.77  reduce.TimeAnyNan2D.time_anynan('float64', (1000, 1000), 'C', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        805±10μs         592±10μs     0.74  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'C', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-        766±10μs         558±40μs     0.73  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'F', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-         708±3μs         510±20μs     0.72  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'F', None, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-        729±30μs         518±20μs     0.71  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'C', None, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-        775±30μs         543±10μs     0.70  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'C', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        801±10μs         540±20μs     0.67  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'C', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-        808±10μs          532±9μs     0.66  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'F', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
       before           after         ratio
     [61eea5ab]       [ff4ee4b4]
     <ppc64le>                  
+        271±10μs         299±20μs     1.10  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'C', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
```